requirements.txt
absl-py==2.2.2
addict==2.4.0
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
alembic==1.14.1
alias-free-torch==0.0.6
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
asteroid-filterbanks==0.4.0
async-timeout==5.0.1
attrs==25.3.0
audioread==3.0.1
bitarray==3.4.2
bitstring==4.3.1
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
coloredlogs==15.0.1
colorlog==6.9.0
contourpy==1.1.1
crcmod==1.7
cryptography==45.0.2
cycler==0.12.1
Cython==3.1.1
datasets==3.1.0
decorator==5.2.1
dill==0.3.8
Distance==0.1.3
dlib==19.24.8
docopt==0.6.2
editdistance==0.8.1
einops==0.8.1
espnet-tts-frontend==0.0.3
et_xmlfile==2.0.0
face-recognition==1.3.0
face_recognition_models==0.3.0
fastcluster==1.2.6
ffmpeg-python==0.2.0
filelock==3.16.1
flatbuffers==25.2.10
fonttools==4.57.0
frozenlist==1.5.0
fsspec==2024.9.0
funasr==1.2.6
future==1.0.0
g2p==2.1.0
g2p-en==2.1.0
google-auth==2.40.2
google-auth-oauthlib==1.0.0
greenlet==3.1.1
grpcio==1.70.0
h5py==3.11.0
hdbscan==0.8.40
huggingface-hub==0.31.4
humanfriendly==10.0
hydra-core==1.3.2
HyperPyYAML==1.2.2
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
inflect==7.0.0
jaconv==0.4.0
jamo==0.4.1
jedi==0.19.2
jieba==0.42.1
Jinja2==3.1.6
jmespath==0.10.0
joblib==1.4.2
julius==0.2.7
kaldiio==2.18.1
kantts==1.0.1
kiwisolver==1.4.7
lazy_loader==0.4
librosa==0.10.1
lightning==2.3.3
lightning-utilities==0.11.9
llvmlite==0.41.1
local-attention==1.10.0
lxml==5.4.0
Mako==1.3.10
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
mdurl==0.1.2
MinDAEC==0.0.2
mir_eval==0.8.2
modelscope==1.26.0
more-itertools==10.5.0
mpmath==1.3.0
ms-funcodec==0.2.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.16
munkres==1.1.4
networkx==3.1
nltk==3.9.1
numba==0.58.1
numpy==1.23.5
oauthlib==3.2.2
omegaconf==2.3.0
onnxruntime==1.19.2
onnxruntime-silicon==1.16.0
openai-whisper @ git+https://github.com/openai/whisper.git@dd985ac4b90cafeef8712f2998d62c59c3e62d22
opencv-python==4.11.0.86
openpyxl==3.1.5
optuna==4.3.0
oss2==2.19.1
packaging==24.2
pandas==2.0.3
panphon==0.20.0
parso==0.8.4
pexpect==4.9.0
phaseaug==1.0.1
pickleshare==0.7.5
pillow==10.4.0
platformdirs==4.3.6
pooch==1.8.2
primePy==1.3
prompt_toolkit==3.0.51
propcache==0.2.0
protobuf==5.29.4
ptflops==0.7.3
ptyprocess==0.7.0
py-sound-connect==0.2.1
pyannote.audio==3.3.0
pyannote.core==5.0.0
pyannote.database==5.1.3
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
Pygments==2.19.1
pynndescent==0.5.13
pyparsing==3.1.4
pypinyin==0.44.0
pysptk==1.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python_speech_features==0.6
pytorch-lightning==2.4.0
pytorch-metric-learning==2.8.1
pytorch-wavelets==1.3.0
pytorch-wpe==0.0.1
pytz==2025.2
PyWavelets==1.4.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rich==14.0.0
rotary-embedding-torch==0.8.6
rsa==4.9.1
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.8
safetensors==0.5.3
scikit-learn==1.0.2
scipy==1.10.1
semver==3.0.4
sentencepiece==0.2.0
shellingham==1.5.4
simplejson==3.20.1
six==1.17.0
sortedcontainers==2.4.0
soundfile==0.13.1
sox==1.5.0
soxr==0.3.7
speechbrain==1.0.3
SQLAlchemy==2.0.41
sympy==1.13.3
tabulate==0.9.0
tensorboard==2.14.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
text-unidecode==1.3
thop==0.1.1.post2209072238
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.20.3
torch==2.4.1
torch-audiomentations==0.12.0
torch-complex==0.4.4
torch_pitch_shift==1.2.5
torchaudio==2.4.1
torchmetrics==1.5.2
torchvision==0.19.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.46.3
ttsfrd==0.1.0
typeguard==2.13.3
typer==0.15.4
typing_extensions==4.13.2
tzdata==2025.2
umap-learn==0.5.7
unicodecsv==0.14.1
Unidecode==1.4.0
urllib3==2.2.3
wcwidth==0.2.13
Werkzeug==3.0.6
whisper==1.1.10
xxhash==3.5.0
yarl==1.15.2
zipp==3.20.2
